+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -9,8 +9,9 @@ import java.util.List;
 import nextstep.core.ApplicationContext;
 import nextstep.mvc.adapter.HandlerAdapter;
 import nextstep.mvc.adapter.RequestMappingHandlerAdapter;
-import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
+import nextstep.mvc.mapping.AnnotationHandlerMapping;
 import nextstep.mvc.exceptionresolver.ExceptionResolverContainer;
+import nextstep.mvc.mapping.HandlerMapping;
 import nextstep.mvc.view.ModelAndView;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
