+++ mvc/src/test/java/nextstep/mvc/controller/tobe/AnnotationHandlerMappingTest.java
@@ -7,9 +7,10 @@ import static org.mockito.Mockito.when;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.core.AnnotationConfigApplicationContext;
-import nextstep.mvc.HandlerMapping;
+import nextstep.mvc.mapping.HandlerMapping;
 import nextstep.mvc.adapter.HandlerAdapter;
 import nextstep.mvc.adapter.RequestMappingHandlerAdapter;
+import nextstep.mvc.mapping.AnnotationHandlerMapping;
 import nextstep.mvc.view.ModelAndView;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
